# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Multi-Agent Software Development System API",
    "version": "0.6.0"
  },
  "paths": {
...
    "/tasks": {
      "get": {
        "summary": "List Tasks",
        "operationId": "list_tasks_tasks_get",
        "tags": [
          "tasks"
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskResponse"
                  },
                  "type": "array",
                  "title": "Response List Tasks Tasks Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Task",
        "operationId": "create_task_tasks_post",
        "tags": [
          "tasks"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}": {
      "get": {
        "summary": "Get Task Detail",
        "operationId": "getTaskDetail",
        "tags": [
          "tasks"
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Task",
        "operationId": "update_task_tasks__task_id__patch",
        "tags": [
          "tasks"
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}/push": {
      "post": {
        "summary": "Push Task To Remote",
        "operationId": "pushTaskToRemote",
        "tags": [
          "tasks"
        ],
        "description": "Pushes the task's work directory to a remote git repository. If git_url is provided in the body, it is saved to the project record before the push.",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/PushTaskRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Push succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushResponse"
                }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Task is not in Completed status; push not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Project has no git_url configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Git push to remote failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/tasks/{task_id}/download": {
      "get": {
        "summary": "Download Task Code",
        "description": "Download the implemented source code as a zip archive.",
        "operationId": "download_task_tasks__task_id__download_get",
        "tags": ["tasks"],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Task Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Zip archive of the working directory",
            "content": { "application/zip": { "schema": { "type": "string", "format": "binary" } } }
          },
          "404": { "description": "Task not found" },
          "409": { "description": "No worker assigned" }
        }
      }
    },
    "/tasks/{task_id}/cleanup": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Initiate Cleanup",
        "description": "Initiates cleanup for a completed or failed task. Transitions to 'cleaning_up' status. Only valid when task status is 'completed' or 'failed'.",
        "operationId": "initiateCleanupTasksTaskIdCleanupPost",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cleanup initiated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CleanupResponse"
                }
              }
            }
          },
          "404": {
            "description": "Task not found"
          },
          "409": {
            "description": "Task is not in a cleanable state (must be completed or failed)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "identifier": {
            "type": "string",
            "title": "Identifier",
            "description": "Machine-readable key; maps to the agent library"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name",
            "description": "Human-readable name shown in the UI"
          }
        },
        "type": "object",
        "required": [
          "id",
          "identifier",
          "display_name"
        ],
        "title": "AgentSummary"
      },
...
      "PushTaskRequest": {
        "properties": {
          "git_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Git Url",
            "description": "Git repository URL to save to the project and push to. If provided, saved to project before push."
          }
        },
        "type": "object",
        "title": "PushTaskRequest"
      },
...
      "ProjectSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "git_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Git Url",
            "description": "Remote git repository URL"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At",
            "description": "Used to disambiguate projects with the same name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source_type"
        ],
        "title": "ProjectSummary"
      },
      "ProjectSummaryWithGitUrl": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "git_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Git Url",
            "description": "Remote git repository URL; used as push target"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source_type"
        ],
        "title": "ProjectSummaryWithGitUrl"
      },
      "TaskStatus": {
        "type": "string",
        "enum": [
          "pending",
          "aborted",
          "in_progress",
          "completed",
          "failed",
          "cleaning_up",
          "cleaned"
        ],
        "title": "TaskStatus"
      },
      "TaskType": {
        "type": "string",
        "enum": [
          "build_feature",
          "fix_bug",
          "review_code",
          "refactor_code",
          "write_tests",
          "scaffold_project"
        ],
        "title": "TaskType"
      },
      "TaskResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "project": {
            "$ref": "#/components/schemas/ProjectSummary"
          },
          "agent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgentSummary"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent",
            "description": "Null for legacy tasks that predate the agent registry"
          },
          "requirements": {
            "type": "string",
            "title": "Requirements"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "task_type": {
            "$ref": "#/components/schemas/TaskType"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project",
          "requirements",
          "status",
          "task_type",
          "created_at",
          "updated_at"
        ],
        "title": "TaskResponse"
      },
      "TaskDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "project": {
            "$ref": "#/components/schemas/ProjectSummaryWithGitUrl"
          },
          "agent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgentSummary"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent",
            "description": "Null for legacy tasks that predate the agent registry"
          },
          "requirements": {
            "type": "string",
            "title": "Requirements"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "work_directory_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Directory Path",
            "description": "Absolute path of the on-disk working directory; null until worker claims the task"
          },
          "elapsed_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elapsed Seconds",
            "description": "Seconds elapsed since started_at; non-null only when status is in_progress"
          },
          "branch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch",
            "default": null
          },
          "assigned_worker_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Worker Id",
            "description": "ID of the worker currently executing or that last executed this task.",
            "default": null
          },
          "assigned_worker_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Worker Url",
            "description": "Internal URL of the worker assigned to this task",
            "default": null
          },
          "task_type": {
            "$ref": "#/components/schemas/TaskType"
          },
          "commits_to_review": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commits To Review",
            "description": "Number of recent commits to review (only for review_code tasks)",
            "default": null
          },
          "required_capabilities": {
            "anyOf": [
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              {
                "type": "null"
              }
            ],
            "title": "Required Capabilities",
            "description": "List of sandbox capabilities required for this task",
            "default": null
          },
          "assigned_sandbox_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Sandbox Id",
            "description": "ID of the sandbox allocated for this task",
            "default": null
          },
          "assigned_sandbox_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Sandbox Url",
            "description": "URL of the sandbox allocated for this task",
            "default": null
          }
        },
        "type": "object",
        "required": [
          "id",
          "project",
          "requirements",
          "status",
          "task_type",
          "created_at",
          "updated_at"
        ],
        "title": "TaskDetailResponse"
      },
      "PushResponse": {
        "properties": {
          "branch_name": {
            "type": "string",
            "title": "Branch Name",
            "description": "Name of the git branch created and pushed to the remote",
            "example": "task/a1b2c3d4"
          },
          "remote_url": {
            "type": "string",
            "title": "Remote Url",
            "description": "The remote git URL that was pushed to"
          },
          "pushed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Pushed At",
            "description": "UTC timestamp of the successful push"
          }
        },
        "type": "object",
        "required": [
          "branch_name",
          "remote_url",
          "pushed_at"
        ],
        "title": "PushResponse"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
      "CreateTaskRequest": {
        "properties": {
          "task_type": {
            "$ref": "#/components/schemas/TaskType"
          },
          "project_name": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Name",
            "description": "Required for scaffold_project tasks. Not required to be unique."
          },
          "agent_id": {
            "type": "string",
            "format": "uuid",
            "title": "Agent Id",
            "description": "ID of the agent from the agent registry. Determines which agent library executes the task."
          },
          "git_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Git Url",
            "description": "Remote git repository URL. Required for all task types except scaffold_project."
          },
          "requirements": {
            "type": "string",
            "minLength": 1,
            "title": "Requirements"
          },
          "branch": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch",
            "description": "Git branch to clone and check out. Required for review_code. Optional for other types. Null = use default branch.",
            "default": null
          },
          "commits_to_review": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commits To Review",
            "description": "Number of recent commits to review. Only valid for review_code tasks.",
            "default": null
          }
        },
        "type": "object",
        "required": [
          "task_type",
          "agent_id",
          "requirements"
        ],
        "title": "CreateTaskRequest"
      },
      "UpdateTaskRequest": {
        "properties": {
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskStatus"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "requirements": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Requirements"
          }
        },
        "type": "object",
        "title": "UpdateTaskRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "CleanupResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "format": "uuid",
            "title": "Task Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "cleaning_up"
            ],
            "title": "Status"
          }
        },
        "required": [
          "task_id",
          "status"
        ],
        "title": "CleanupResponse",
        "type": "object"
      },
...
    }
  }
}